+++ finchge/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.0.1-beta.14"
+__version__ = "1.0.1-beta.15"
 
 from typing import TYPE_CHECKING, Any